flask_backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SSE==1.0.0
celery==5.4.0
redis==5.2.1
primer3-py==2.0.3
pydantic==2.10.6
numpy==1.23.5
biopython==1.81
prettytable==3.12.0
tqdm==4.67.1
watchdog==3.0.0
pyyaml==6.0
